move some space ops to separate package
Use a separate package for this so that it can be imported by both the
apiserver and oldconfig package.
